+++ src/players/base_player.py
@@ -26,7 +26,7 @@ class BasePlayer(ABC):
         self,
         state: np.ndarray,
         action: int,
-        reward: int,
+        reward: float,
         next_state: np.ndarray,
         terminal: bool,
     ) -> None:
